HSEARCH-2772 Move AWS Elasticsearch request signing to a dedicated module
