+++ elasticsearch-aws/src/main/resources/META-INF/services/org.hibernate.search.elasticsearch.client.spi.ElasticsearchHttpClientConfigurer
@@ -1,0 +1,1 @@
+org.hibernate.search.elasticsearch.aws.impl.AWSElasticsearchHttpClientConfigurer
