--- integrationtest/elasticsearch/src/test/resources/META-INF/services/org.hibernate.search.elasticsearch.client.spi.ElasticsearchHttpClientConfigurer
@@ -1,1 +1,0 @@
-org.hibernate.search.elasticsearch.testutil.aws.AWSElasticsearchHttpClientConfigurer
